Expose Placemark.properties
Also refactored GeocodedPlacemark out of Placemark to improve
readability.
