+++ MapboxGeocoder/MBPlacemark.swift
@@ -87,14 +87,13 @@ open class Placemark: NSObject, Codable {
         name = try container.decode(String.self, forKey: .name)
         address = try container.decodeIfPresent(String.self, forKey: .address)
         qualifiedName = try container.decodeIfPresent(String.self, forKey: .qualifiedName)
-        superiorPlacemarks = try container.decodeIfPresent([GeocodedPlacemark].self, forKey: .superiorPlacemarks)
+        superiorPlacemarks = try container.decodeIfPresent([Placemark].self, forKey: .superiorPlacemarks)
         
         if let coordinates = try container.decodeIfPresent([CLLocationDegrees].self, forKey: .centerCoordinate) {
             let coordinate = CLLocationCoordinate2D(geoJSON: coordinates)
             location = CLLocation(coordinate: coordinate)
         }
         
-        code = try container.decodeIfPresent(String.self, forKey: .code)?.uppercased()
         if let rawIdentifier = try container.decodeIfPresent(String.self, forKey: .wikidataItemIdentifier) {
             let identifier = rawIdentifier.trimmingCharacters(in: .whitespacesAndNewlines)
             assert(identifier.hasPrefix("Q"))
@@ -102,6 +101,7 @@ open class Placemark: NSObject, Codable {
         }
         
         properties = try container.decodeIfPresent(Properties.self, forKey: .properties)
+        code = try container.decodeIfPresent(String.self, forKey: .code)?.uppercased()
         
         if let boundingBox = try container.decodeIfPresent([CLLocationDegrees].self, forKey: .boundingBox) {
             let southWest = CLLocationCoordinate2D(geoJSON: Array(boundingBox.prefix(2)))
@@ -152,7 +152,7 @@ open class Placemark: NSObject, Codable {
     /**
      A subset of the `properties` object on a GeoJSON feature suited for Geocoding results.
      */
-    fileprivate var properties: Properties?
+    public var properties: Properties?
     
     /**
      The common name of the placemark.
@@ -228,15 +228,6 @@ open class Placemark: NSObject, Codable {
     @objc open var region: CLRegion?
     
     // MARK: Accessing Contact Information
-    
-    /**
-     The placemark’s full address in the customary local format, with each line in a separate string in the array.
-     
-     If you need to fit the same address on a single line, use the `qualifiedName` property, in which each line is separated by a comma instead of a line break.
-     */
-    fileprivate var formattedAddressLines: [String]? {
-        return nil
-    }
     
     #if !os(tvOS)
     /**
@@ -367,22 +358,41 @@ internal struct GeocodeResult: Codable {
 }
 
 /**
- A subset of the `properties` object on a GeoJSON feature suited for Geocoding results.
+ An object describing the placemark. Only Carmen (https://github.com/mapbox/carmen) properties are guaranteed.
  */
-internal struct Properties: Codable {
+public class Properties: Codable {
     private enum CodingKeys: String, CodingKey {
         case shortCode = "short_code"
         case phoneNumber = "tel"
         case maki
         case address
         case category
+        case landmark
+        case wikidata
     }
     
-    let shortCode: String?
-    let maki: String?
-    let phoneNumber: String?
-    let address: String?
-    let category: String?
+    // The ISO 3166-1 country and ISO 3166-2 region code for the returned feature.
+    public let shortCode: String?
+    
+    // The name of a suggested Maki icon (https://www.mapbox.com/maki-icons/) to visualize a  poi feature based on its category.
+    public let maki: String?
+    
+    // A formatted string of the telephone number for the returned  poi feature.
+    public let phoneNumber: String?
+    
+    // A string of the full street address for the returned  poi feature.
+    // Note that unlike the  address property for  address features, this property is inside the properties object.
+    public let address: String?
+    
+    // A boolean value indicating whether a  poi feature is a landmark.
+    // Landmarks are particularly notable or long-lived features like schools, parks, museums and places of worship.
+    public let landmark: Bool?
+    
+    // A string of comma-separated categories for the returned  poi feature.
+    public let category: String?
+    
+    // The Wikidata identifier for the returned feature.
+    public let wikidata: String?
 }
 
 // Used internally for flattening and transforming routable_points.points.coordinates
@@ -394,188 +404,6 @@ internal struct RoutableLocation: Codable {
             return CLLocationCoordinate2D(latitude: coordinates[1], longitude: coordinates[0])
         }
         return nil
-    }
-}
-
-/**
- A concrete subclass of `Placemark` to represent results of geocoding requests.
- */
-@objc(MBGeocodedPlacemark)
-open class GeocodedPlacemark: Placemark {
-    
-    private enum CodingKeys: String, CodingKey {
-        case routableLocations = "routable_points"
-        case relevance
-    }
-    
-    private enum PointsCodingKeys: String, CodingKey {
-        case points
-    }
-    
-    /**
-     An array of locations that serve as hints for navigating to the placemark.
-     
-     If the `GeocodeOptions.includesRoutableLocations` property is set to `true`, this property contains locations that are suitable to use as a waypoint in a routing engine such as MapboxDirections.swift. Otherwise, if the `GeocodeOptions.includesRoutableLocations` property is set to `false`, this property is set to `nil`.
-     
-     For the placemark’s geographic center, use the `location` property. The routable locations may differ from the geographic center. For example, if a house’s driveway leads to a street other than the nearest street (by straight-line distance), then this property may contain the location where the driveway meets the street. A route to the placemark’s geographic center may be impassable, but a route to the routable location would end on the correct street with access to the house.
-     */
-    @objc open var routableLocations: [CLLocation]?
-    
-    public required init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        if let pointsContainer = try? container.nestedContainer(keyedBy: PointsCodingKeys.self, forKey: .routableLocations),
-            var coordinatesContainer = try? pointsContainer.nestedUnkeyedContainer(forKey: .points) {
-            
-            if let routableLocation = try coordinatesContainer.decodeIfPresent(RoutableLocation.self),
-                let coordinate = routableLocation.coordinate {
-                routableLocations = [CLLocation(coordinate: coordinate)]
-            }
-        }
-        
-        relevance = try container.decodeIfPresent(Double.self, forKey: .relevance) ?? -1
-        
-        try super.init(from: decoder)
-    }
-    
-    public override func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        try container.encodeIfPresent(relevance, forKey: .relevance)
-        
-        if let routableLocations = routableLocations,
-            !routableLocations.isEmpty {
-            var pointsContainer = container.nestedContainer(keyedBy: PointsCodingKeys.self, forKey: .routableLocations)
-            var coordinatesContainer = pointsContainer.nestedUnkeyedContainer(forKey: .points)
-            let routableLocation = RoutableLocation(coordinates: [routableLocations[0].coordinate.longitude,
-                                                                  routableLocations[0].coordinate.latitude])
-            try coordinatesContainer.encode(routableLocation)
-        }
-        
-        try super.encode(to: encoder)
-    }
-    
-    @objc open override var debugDescription: String {
-        return qualifiedName!
-    }
-    
-    internal var qualifiedNameComponents: [String] {
-        if qualifiedName!.contains(", ") {
-            return qualifiedName!.components(separatedBy: ", ")
-        }
-        // Chinese addresses have no commas and are reversed.
-        return (superiorPlacemarks?.map { $0.name } ?? []).reversed() + [name]
-    }
-    
-    @objc open var formattedName: String {
-        let text = super.name
-        // For address features, `text` is just the street name. Look through the fully-qualified address to determine whether to put the house number before or after the street name.
-        if let houseNumber = address, scope == .address {
-            let streetName = text
-            let reversedAddress = "\(streetName) \(houseNumber)"
-            if qualifiedNameComponents.contains(reversedAddress) {
-                return reversedAddress
-            } else {
-                return "\(houseNumber) \(streetName)"
-            }
-        } else {
-            return text
-        }
-    }
-    
-    @objc open override var genres: [String]? {
-        return properties?.category?.components(separatedBy: ", ")
-    }
-    
-    @objc open override var imageName: String? {
-        return properties?.maki
-    }
-    
-    /**
-     A numerical score from 0 (least relevant) to 0.99 (most relevant) measuring
-     how well each returned feature matches the query. Use this property to
-     remove results that don’t fully match the query.
-     */
-    @objc open var relevance: Double
-    
-    private var clippedAddressLines: [String] {
-        let lines = qualifiedNameComponents
-        if scope == .address {
-            return lines
-        }
-        
-        guard let qualifiedName = qualifiedName,
-            qualifiedName.contains(", ") else {
-            // Chinese addresses have no commas and are reversed.
-            return Array(lines.prefix(lines.count))
-        }
-        
-        return Array(lines.suffix(from: 1))
-    }
-    
-    override var formattedAddressLines: [String] {
-        return clippedAddressLines
-    }
-    
-    #if !os(tvOS)
-    @available(iOS 9.0, OSX 10.11, *)
-    @objc open override var postalAddress: CNPostalAddress? {
-        let postalAddress = CNMutablePostalAddress()
-        
-        if scope == .address {
-            postalAddress.street = name
-        } else if let address = address {
-            postalAddress.street = address.replacingOccurrences(of: ", ", with: "\n")
-        }
-
-        if let placeName = place?.name {
-            postalAddress.city = placeName
-        }
-        if let regionName = administrativeRegion?.name {
-            postalAddress.state = regionName
-        }
-        if let postalCode = postalCode?.name {
-            postalAddress.postalCode = postalCode
-        }
-        if let countryName = country?.name {
-            postalAddress.country = countryName
-        }
-        if let ISOCountryCode = country?.code {
-            postalAddress.isoCountryCode = ISOCountryCode
-        }
-    
-        return postalAddress
-    }
-    #endif
-    
-    @objc open override var addressDictionary: [AnyHashable: Any]? {
-        var addressDictionary: [String: Any] = [:]
-        if scope == .address {
-            addressDictionary[MBPostalAddressStreetKey] = name
-        } else if let address = properties?.address {
-            addressDictionary[MBPostalAddressStreetKey] = address
-        } else if let address = address {
-            addressDictionary[MBPostalAddressStreetKey] = address
-        }
-        addressDictionary[MBPostalAddressCityKey] = place?.name
-        addressDictionary[MBPostalAddressStateKey] = administrativeRegion?.name
-        addressDictionary[MBPostalAddressPostalCodeKey] = postalCode?.name
-        addressDictionary[MBPostalAddressCountryKey] = country?.name
-        addressDictionary[MBPostalAddressISOCountryCodeKey] = country?.code
-        addressDictionary["formattedAddressLines"] = clippedAddressLines
-        addressDictionary["name"] = name
-        addressDictionary["subAdministrativeArea"] = district?.name ?? place?.name
-        addressDictionary["subLocality"] = neighborhood?.name
-        addressDictionary["subThoroughfare"] = subThoroughfare
-        addressDictionary["thoroughfare"] = thoroughfare
-        return addressDictionary
-    }
-    
-    /**
-     The phone number to contact a business at this location.
-     */
-    @objc open override var phoneNumber: String? {
-        return properties?.phoneNumber
     }
 }
