+++ MapboxGeocoderTests/fixtures/forward_valid_hk.json
@@ -1,0 +1,1 @@
+{"type":"FeatureCollection","query":["hong","kong"],"features":[{"id":"place.364","type":"Feature","place_type":["country","region","place"],"relevance":1,"properties":{"wikidata":"Q8646","short_code":"hk"},"text":"Hong Kong","place_name":"Hong Kong","bbox":[113.7734944,22.134324,114.5045213,22.573549],"center":[114.135442,22.346578],"geometry":{"type":"Point","coordinates":[114.135442,22.346578]}},{"id":"place.12997319628474470","type":"Feature","place_type":["place"],"relevance":1,"properties":{"wikidata":"Q8646"},"text":"Hong Kong","place_name":"Hong Kong, Hong Kong","bbox":[114.080826,22.198214,114.249146,22.321415],"center":[114.164054,22.25959],"geometry":{"type":"Point","coordinates":[114.164054,22.25959]},"context":[{"id":"region.364","short_code":"hk","wikidata":"Q8646","text":"Hong Kong"}]},{"id":"poi.13251169","type":"Feature","place_type":["poi"],"relevance":1,"properties":{"tel":"2721 0116","address":"10 Salisbury Rd","category":"museum","landmark":true,"wikidata":"Q908216","maki":"museum"},"text":"Hong Kong Museum of Art","place_name":"Hong Kong Museum of Art, 10 Salisbury Rd, Kowloon, Hong Kong","center":[114.17208,22.293325],"geometry":{"type":"Point","coordinates":[114.17208,22.293325]},"context":[{"id":"locality.4068150428060122","wikidata":"Q157669","text":"Youjianwang Qu"},{"id":"place.8363024200553460","wikidata":"Q239143","text":"Kowloon"},{"id":"region.364","short_code":"hk","wikidata":"Q8646","text":"Hong Kong"}]},{"id":"poi.13242214","type":"Feature","place_type":["poi"],"relevance":1,"properties":{"tel":"2732 3232","address":"2 Science Museum Rd","category":"museum","landmark":true,"wikidata":"Q836319","maki":"museum"},"text":"Hong Kong Science Museum","place_name":"Hong Kong Science Museum, 2 Science Museum Rd, Kowloon, Hong Kong","center":[114.17726,22.301245],"geometry":{"type":"Point","coordinates":[114.17726,22.301245]},"context":[{"id":"locality.4068150428060122","wikidata":"Q157669","text":"Youjianwang Qu"},{"id":"place.8363024200553460","wikidata":"Q239143","text":"Kowloon"},{"id":"region.364","short_code":"hk","wikidata":"Q8646","text":"Hong Kong"}]},{"id":"poi.13245629","type":"Feature","place_type":["poi"],"relevance":1,"properties":{"tel":"2355 7234","address":"9 Cheong Wan Rd","category":"music, show venue, concert, concert hall","landmark":true,"wikidata":"Q835531","maki":"music"},"text":"Hong Kong Coliseum","place_name":"Hong Kong Coliseum, 9 Cheong Wan Rd, Kowloon, Hong Kong","center":[114.18203,22.301416],"geometry":{"type":"Point","coordinates":[114.18203,22.301416]},"context":[{"id":"locality.4068150428060122","wikidata":"Q157669","text":"Youjianwang Qu"},{"id":"place.8363024200553460","wikidata":"Q239143","text":"Kowloon"},{"id":"region.364","short_code":"hk","wikidata":"Q8646","text":"Hong Kong"}]},{"id":"poi.2411473","type":"Feature","place_type":["poi"],"relevance":0.99,"properties":{"address":"16th St","category":"stadium, arena","landmark":true,"wikidata":"Q7440659","maki":"baseball"},"text":"Seals Stadium","place_name":"Seals Stadium, 16th St, San Francisco, California 94103, United States","matching_text":"stadium","matching_place_name":"stadium, 16th St, San Francisco, California 94103, United States","center":[-122.41035,37.76573],"geometry":{"type":"Point","coordinates":[-122.41035,37.76573]},"context":[{"id":"neighborhood.293934","text":"Mission"},{"id":"postcode.13892342768265050","text":"94103"},{"id":"place.15734669613361910","wikidata":"Q62","text":"San Francisco"},{"id":"region.3591","short_code":"US-CA","wikidata":"Q99","text":"California"},{"id":"country.3145","short_code":"us","wikidata":"Q30","text":"United States"}]}],"attribution":"NOTICE: © 2018 Mapbox and its suppliers. All rights reserved. Use of this data is subject to the Mapbox Terms of Service (https://www.mapbox.com/about/maps/). This response and the information it contains may not be retained."}
